chore(app): extract toFlagNames outside the app component
