+++ src/containers/toFlagNames.js
@@ -1,0 +1,15 @@
+import isObjectLike from 'lodash/isObjectLike'
+import isArray from 'lodash/isArray'
+import keys from 'lodash/keys'
+import flatten from 'lodash/flatten'
+
+// TODO add this to cozy-flags ?
+export const toFlagNames = (flagName, prefix = '') => {
+  if (typeof flagName === 'string') return `${prefix}${flagName}`
+  else if (isArray(flagName))
+    return flatten(flagName.map(flagName => toFlagNames(flagName, prefix)))
+  else if (isObjectLike(flagName))
+    return flatten(
+      keys(flagName).map(key => toFlagNames(flagName[key], `${prefix}${key}.`))
+    )
+}
